# backend/openapi/spec.yaml
openapi: "3.0.3"
info:
  title: Microcord API
  version: "0.9.2"
  description: Minimal Discord-like self-hosted chat and voice API.
servers:
  - url: /api
paths:
...
  /auth/refresh:
    post:
      operationId: api.auth.refresh
      summary: Rotate a refresh token to get a new access/refresh token pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "401":
          description: Invalid or expired refresh token
        "429":
          description: Rate limited
...
  /auth/ws-ticket:
    post:
      operationId: api.auth.ws_ticket
      summary: Issue a one-time WebSocket ticket (requires JWT auth)
      responses:
        "200":
          description: Ticket issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket:
                    type: string
        "401":
          description: Not authenticated
...
  /users/{user_id}:
    get:
      operationId: api.users.get_user
      summary: Get a user by ID
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found

    patch:
      operationId: api.users.update_user
      summary: Update own profile
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          description: Cannot modify another user
        "404":
          description: User not found
  /users/{user_id}/admin:
    post:
      operationId: api.users.set_user_admin
      summary: Promote or demote a user's admin status (admin/owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetAdminRequest"
      responses:
        "200":
          description: Admin status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          description: Not authorized or cannot modify owner
        "404":
          description: User not found
  /users/{user_id}/recover:
    post:
      operationId: api.users.recover_account
      summary: Initiate account recovery for a user (owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Recovery passphrase generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryResponse"
        "403":
          description: Not owner
        "404":
          description: User not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: services.auth.decode_token
  schemas:
...
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: The raw refresh token to rotate
    RefreshResponse:
      type: object
      properties:
        access_token:
          type: string
          description: New short-lived JWT access token
        refresh_token:
          type: string
          description: New long-lived opaque refresh token
    UpdateUserRequest:
      type: object
      additionalProperties: false
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 40
        tick_sound:
          type: integer
          minimum: 1
          maximum: 4
    SetAdminRequest:
      type: object
      additionalProperties: false
      required:
        - is_admin
      properties:
        is_admin:
          type: boolean
    RecoveryResponse:
      type: object
      properties:
        recovery_passphrase:
          type: string
          description: One-time recovery passphrase (shown only once)
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: Null if the passphrase never expires (owner's own account)
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
          nullable: true
        tick_sound:
          type: integer
          minimum: 1
          maximum: 4
        is_admin:
          type: boolean
        is_owner:
          type: boolean
        created_at:
          type: string
          format: date-time
...